+++ src/containers/WebComponentLoader.test.js
@@ -1,4 +1,4 @@
-import { render, act } from "@testing-library/react";
+import { render, act, screen } from "@testing-library/react";
 import React from "react";
 import { Provider } from "react-redux";
 import configureStore from "redux-mock-store";
@@ -663,5 +663,41 @@ describe("When user is in state", () => {
         }),
       );
     });
+  });
+});
+
+describe("When isLoading prop is used", () => {
+  beforeEach(() => {
+    const middlewares = [];
+    const mockStore = configureStore(middlewares);
+    const initialState = {
+      editor: {
+        loading: "success",
+        project: {
+          components: [],
+        },
+        openFiles: [],
+        focussedFileIndices: [],
+        hasShownSavePrompt: false,
+        remixLoadFailed: false,
+        justLoaded: false,
+        saveTriggered: false,
+      },
+      instructions: {},
+      auth: {},
+    };
+    store = mockStore(initialState);
+    cookies = new Cookies();
+  });
+
+  test("it respects isLoading prop when rendering", () => {
+    render(
+      <Provider store={store}>
+        <CookiesProvider cookies={cookies}>
+          <WebComponentLoader isLoading={true} />
+        </CookiesProvider>
+      </Provider>,
+    );
+    expect(screen.queryByText("webComponent.loading")).toBeInTheDocument();
   });
 });